courses 앱 분리
