+++ apps/courses/models/__init__.py
@@ -1,0 +1,4 @@
+from apps.courses.models.cohort import Cohort
+from apps.courses.models.subject import Subject
+
+__all__ = ["Cohort", "Subject"]
