+++ apps/posts/models/__init__.py
@@ -1,9 +1,8 @@
+from apps.courses.models import Cohort, Subject
 from apps.posts.models.category import PostCategory
-from apps.posts.models.cohort import *
 from apps.posts.models.comment import PostComment, PostCommentTag
 from apps.posts.models.course import Course
 from apps.posts.models.like import Like
 from apps.posts.models.post import Post
-from apps.posts.models.subject import Subject
 
 __all__ = ["PostCategory", "Post", "PostComment", "PostCommentTag", "Like", "Subject", "Course", "Cohort"]
