+++ apps/users/tests/user_info_test.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from apps.posts.models.cohort import Cohort
+from apps.courses.models.cohort import Cohort
 from apps.posts.models.course import Course
 from apps.users.models import CohortStudents, User
